requirements.txt

PyYAML>=5.1
SQLAlchemy>=1.3
tqdm>=4.42
tweepy>=3.7.0
click>=7.1.2

requests~=2.24.0
pymongo~=3.10.1